+++ Packages/com.mimylab.fukuroudon/Runtime/ActiveRelay/Scripts/ActiveRelayByPickup.cs
@@ -8,7 +8,7 @@ namespace MimyLab
 {
     using UdonSharp;
     using UnityEngine;
-    //using VRC.SDKBase;
+    using VRC.SDKBase;
     //using VRC.Udon;
     //using VRC.SDK3.Components;
 
@@ -21,17 +21,53 @@ namespace MimyLab
     }
 
     [AddComponentMenu("Fukuro Udon/Active Relay/ActiveRelay by Pickup")]
-    [UdonBehaviourSyncMode(BehaviourSyncMode.None)]
+    [UdonBehaviourSyncMode(BehaviourSyncMode.Any)]
     public class ActiveRelayByPickup : ActiveRelayBy
     {
         [SerializeField]
         private ActiveRelayPickupEventType _eventType = default;
+        [SerializeField]
+        private bool _localOnly = true;
+
+        [UdonSynced]
+        private bool[] sync_objectsActive = new bool[0];
+
+        private void Start()
+        {
+            sync_objectsActive = new bool[_gameObjects.Length];
+            for (int i = 0; i < sync_objectsActive.Length; i++)
+            {
+                sync_objectsActive[i] = _gameObjects[i].activeSelf;
+            }
+        }
+
+        public override void OnPreSerialization()
+        {
+            for (int i = 0; i < sync_objectsActive.Length; i++)
+            {
+                sync_objectsActive[i] = _gameObjects[i].activeSelf;
+            }
+        }
+
+        public override void OnDeserialization()
+        {
+            if (_localOnly) { return; }
+
+            for (int i = 0; i < _gameObjects.Length; i++)
+            {
+                if (_gameObjects[i].activeSelf != sync_objectsActive[i])
+                {
+                    _gameObjects[i].SetActive(sync_objectsActive[i]);
+                }
+            }
+        }
 
         public override void OnPickup()
         {
             if (_eventType == ActiveRelayPickupEventType.Pickup)
             {
                 DoAction();
+                Sync();
             }
         }
 
@@ -40,6 +76,7 @@ namespace MimyLab
             if (_eventType == ActiveRelayPickupEventType.PickupUseDown)
             {
                 DoAction();
+                Sync();
             }
         }
 
@@ -48,6 +85,7 @@ namespace MimyLab
             if (_eventType == ActiveRelayPickupEventType.PickupUseUp)
             {
                 DoAction();
+                Sync();
             }
         }
 
@@ -56,7 +94,19 @@ namespace MimyLab
             if (_eventType == ActiveRelayPickupEventType.Drop)
             {
                 DoAction();
+                Sync();
             }
+        }
+
+        private void Sync()
+        {
+            if (_localOnly) { return; }
+
+            if (!Networking.IsOwner(this.gameObject))
+            {
+                Networking.SetOwner(Networking.LocalPlayer, this.gameObject);
+            }
+            RequestSerialization();
         }
     }
 }